refactor(benchmark): change type of loop_count (int -> size_t)

+++ tests/stack/StackBenchmark.cpp
@@ -1,10 +1,10 @@
 #include "stack/StackBenchmark.hpp"
 
-const int StackBenchmark::loop_count = 10000;
+const size_t StackBenchmark::loop_count = 10000;
 
 StackBenchmark::StackBenchmark() : timer(Timer()), original_vec()
 {
-    for (int i = 0; i < loop_count; i++)
+    for (size_t i = 0; i < loop_count; i++)
     {
         original_vec.push_back(i);
     }
@@ -21,7 +21,7 @@ void StackBenchmark::test_top()
 {
     ft::stack<int, ft::vector<int> > stk(original_vec);
     timer.start();
-    for (int i = 0; i < loop_count; i++)
+    for (size_t i = 0; i < loop_count; i++)
     {
         stk.top();
     }
@@ -34,7 +34,7 @@ void StackBenchmark::test_push()
     ft::vector<int> empty_vec;
     ft::stack<int, ft::vector<int> > stk(empty_vec);
     timer.start();
-    for (int i = 0; i < loop_count; i++)
+    for (size_t i = 0; i < loop_count; i++)
     {
         stk.push(i);
     }
@@ -46,7 +46,7 @@ void StackBenchmark::test_pop()
 {
     ft::stack<int, ft::vector<int> > stk(original_vec);
     timer.start();
-    for (int i = 0; i < loop_count; i++)
+    for (size_t i = 0; i < loop_count; i++)
     {
         stk.pop();
     }
